reorganize so tests run
